+++ playground/nextjs/pages/index.tsx
@@ -57,8 +57,10 @@ export default function Home() {
                 <div className="buttons">
                     <Link href="/animations">Animations</Link>
                     <Link href="/iframe">Iframe</Link>
+                    <Link href="/canvas">Canvas</Link>
                     <Link href="/media">Media</Link>
                     <Link href="/long">Long</Link>
+                    <Link href="/longmain">Long Main</Link>
                 </div>
 
                 <p>Feature flag response: {JSON.stringify(result)}</p>