+++ src/__tests__/posthog-persistence.test.ts
@@ -23,7 +23,8 @@ describe('persistence', () => {
         referrer = ''
     })
 
-    describe.each([`cookie`, `localStorage`, `localStorage+cookie`])('persistence modes: %p', (persistenceMode) => {
+    const persistenceModes: string[] = ['cookie', 'localStorage', 'localStorage+cookie']
+    describe.each(persistenceModes)('persistence modes: %p', (persistenceMode) => {
         // Common tests for all storage modes
         beforeEach(() => {
             library = new PostHogPersistence(makePostHogConfig('test', persistenceMode))
@@ -115,12 +116,12 @@ describe('persistence', () => {
             const lib = new PostHogPersistence(makePostHogConfig('bla', 'cookie'))
             lib.register_once({ distinct_id: 'testy', test_prop: 'test_value' }, undefined, undefined)
             expect(document.cookie).toEqual(
-                '; ph__posthog=%7B%22distinct_id%22%3A%22testy%22%2C%22test_prop%22%3A%22test_value%22%7D'
+                'ph__posthog=%7B%22distinct_id%22%3A%22testy%22%2C%22test_prop%22%3A%22test_value%22%7D'
             )
             const lib2 = new PostHogPersistence(makePostHogConfig('bla', 'localStorage+cookie'))
-            expect(document.cookie).toEqual('; ph__posthog=%7B%22distinct_id%22%3A%22testy%22%7D')
+            expect(document.cookie).toEqual('ph__posthog=%7B%22distinct_id%22%3A%22testy%22%7D')
             lib2.register({ test_prop2: 'test_val', distinct_id: 'test2' })
-            expect(document.cookie).toEqual('; ph__posthog=%7B%22distinct_id%22%3A%22test2%22%7D')
+            expect(document.cookie).toEqual('ph__posthog=%7B%22distinct_id%22%3A%22test2%22%7D')
             expect(lib2.props).toEqual({ distinct_id: 'test2', test_prop: 'test_value', test_prop2: 'test_val' })
             lib2.remove()
             expect(localStorage.getItem('ph__posthog')).toEqual(null)